Add workaround getting spaces through shell
Sometimes you do weird stuff and it's tough to get spaces properly
quoted in the final shell command. This option lets you replace spaces
with dots so you don't need to worry about it (unless your searcher
doesn't support regex).

This came up because I'm trying to use grep with a filelist and am doing
so with xargs.