Accept relative paths to UseGrepRecursiveFrom
Also fix UseGrepRecursiveFrom on macOS (probably Linux too).

Expand path so we can accept relative paths, but always search from the
same directory regardless of where we are in the project.

Need to add quotes around $* for any queries (regardless of path) to
work from macOS. Not sure if it's a difference in grep's unixiness or
macvim.