Convert \c and \C case-matching atom to PCRE
I don't see any way to do \C -- only Tcl supports (?c). So omit it.
Would be nice to pass the appropriate flag to the prg.

Note that grep doesn't support (?c) without -P (--perl-regexp), but it's
not really intended for you to use grep with notgrep. If you're doing
that, then you need to sanitize the Perl regex yourself. I don't see a
pretty way of doing it.